Move constituencies view to elections app, rename
